Компонент header, рефакторинг компонентов
