+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,9 +1,7 @@
 import {Offer} from '../../types/offer';
-import {AppRoute} from '../../const';
-import {Link} from 'react-router-dom';
 import {useAppSelector} from '../../hooks';
-import Logo from '../../components/logo/logo';
 import FavoritesList from '../../components/favorites-list/favorites-list';
+import Header from '../../components/header/header';
 
 export default function FavoritesScreen(): JSX.Element {
   const offers = useAppSelector((state) => state.offers);
@@ -11,34 +9,7 @@ export default function FavoritesScreen(): JSX.Element {
 
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo />
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="/">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <Link to={AppRoute.Favorites}>
-                      <span className="header__favorite-count">3</span>
-                    </Link>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="/">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header />
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
