+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,19 +1,17 @@
 import {useState} from 'react';
-import {Link} from 'react-router-dom';
 import {Offer} from '../../types/offer';
-import {AppRoute, SortOption, AuthorizationStatus} from '../../const';
-import Logo from '../../components/logo/logo';
+import {SortOption} from '../../const';
 import PlacesList from '../../components/places-list/places-list';
 import CitiesList from '../../components/cities-list/cities-list';
 import SortOptions from '../../components/sort-options/sort-options';
 import Map from '../../components/map/map';
 import {useAppSelector} from '../../hooks';
+import Header from '../../components/header/header';
 
 export default function MainScreen(): JSX.Element {
   const additionalMapClass = 'cities__map';
   const placesType = 'cities';
 
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const currentCity = useAppSelector((state) => state.city);
   const currentOffers = useAppSelector((state) => state.offers).filter((offer) => offer.city.name === currentCity);
   const currentCityInfo = currentOffers[0].city;
@@ -46,45 +44,7 @@ export default function MainScreen(): JSX.Element {
 
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo />
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                {authorizationStatus === AuthorizationStatus.Authorized &&
-                <>
-                  <li className="header__nav-item user">
-                    <a className="header__nav-link header__nav-link--profile" href="/">
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                      </div>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                      <Link to={AppRoute.Favorites}>
-                        <span className="header__favorite-count">3</span>
-                      </Link>
-                    </a>
-                  </li>
-                  <li className="header__nav-item">
-                    <a className="header__nav-link" href="/">
-                      <span className="header__signout">Sign out</span>
-                    </a>
-                  </li>
-                </>}
-                {authorizationStatus === AuthorizationStatus.NotAuthorized &&
-                  <li className="header__nav-item user">
-                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                      </div>
-                      <span className="header__login">Sign in</span>
-                    </Link>
-                  </li>}
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header />
 
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
